# docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GW2 WvW日志系统",
    "description": "GW2 WvW日志解析与评分系统后端API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/dictionary/types": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取字典类型列表",
        "description": "获取所有字典类型，支持分页和状态筛选",
        "operationId": "get_dict_types_api_v1_dictionary_types_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态筛选：0-启用，1-禁用",
              "title": "Status"
            },
            "description": "状态筛选：0-启用，1-禁用"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键词筛选：按名称或编码搜索",
              "title": "Keyword"
            },
            "description": "关键词筛选：按名称或编码搜索"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "字典管理"
        ],
        "summary": "创建字典类型",
        "description": "创建新的字典类型",
        "operationId": "create_dict_type_api_v1_dictionary_types_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictTypeCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/types/all": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取所有启用的字典类型",
        "description": "获取所有启用的字典类型，不分页",
        "operationId": "get_all_dict_types_api_v1_dictionary_types_all_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/types/{dict_id}": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取单个字典类型",
        "description": "获取指定的字典类型详情",
        "operationId": "get_dict_type_api_v1_dictionary_types__dict_id__get",
        "parameters": [
          {
            "name": "dict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "字典管理"
        ],
        "summary": "更新字典类型",
        "description": "更新字典类型信息",
        "operationId": "update_dict_type_api_v1_dictionary_types__dict_id__put",
        "parameters": [
          {
            "name": "dict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "字典管理"
        ],
        "summary": "删除字典类型",
        "description": "删除字典类型及关联的字典项",
        "operationId": "delete_dict_type_api_v1_dictionary_types__dict_id__delete",
        "parameters": [
          {
            "name": "dict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/data": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取字典项列表",
        "description": "获取字典项列表，支持按字典类型筛选",
        "operationId": "get_dict_data_api_v1_dictionary_data_get",
        "parameters": [
          {
            "name": "dict_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "字典类型编码",
              "title": "Dict Type"
            },
            "description": "字典类型编码"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "每页数量",
              "default": 50,
              "title": "Page Size"
            },
            "description": "每页数量"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态筛选",
              "title": "Status"
            },
            "description": "状态筛选"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "字典管理"
        ],
        "summary": "创建字典项",
        "description": "创建新的字典项",
        "operationId": "create_dict_data_api_v1_dictionary_data_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictDataCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/data/{dict_code}": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取单个字典项",
        "description": "获取指定的字典项详情",
        "operationId": "get_dict_data_detail_api_v1_dictionary_data__dict_code__get",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "字典管理"
        ],
        "summary": "更新字典项",
        "description": "更新字典项信息",
        "operationId": "update_dict_data_api_v1_dictionary_data__dict_code__put",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictDataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "字典管理"
        ],
        "summary": "删除字典项",
        "description": "删除指定的字典项",
        "operationId": "delete_dict_data_api_v1_dictionary_data__dict_code__delete",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dict Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/options/{dict_type}": {
      "get": {
        "tags": [
          "字典管理"
        ],
        "summary": "获取字典选项",
        "description": "获取指定字典类型的下拉选项（公开接口）",
        "operationId": "get_dict_options_api_v1_dictionary_options__dict_type__get",
        "parameters": [
          {
            "name": "dict_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dict Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/reload-cache": {
      "post": {
        "tags": [
          "字典管理"
        ],
        "summary": "刷新字典缓存",
        "description": "手动刷新所有字典数据到内存缓存",
        "operationId": "reload_dict_cache_api_v1_dictionary_reload_cache_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/init": {
      "post": {
        "tags": [
          "字典管理"
        ],
        "summary": "初始化字典数据",
        "description": "初始化系统预置字典数据",
        "operationId": "init_dictionary_data_api_v1_dictionary_init_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "操作是否成功"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示消息"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "错误代码"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "时间戳"
          },
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "HTTP状态码"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ApiResponse"
      },
...
      "DictTypeCreate": {
        "properties": {
          "dict_type": {
            "type": "string",
            "title": "Dict Type",
            "description": "字典类型编码"
          },
          "dict_name": {
            "type": "string",
            "title": "Dict Name",
            "description": "字典类型名称"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "状态：0-启用，1-禁用",
            "default": 0
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "description": "排序顺序",
            "default": 0
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注说明"
          }
        },
        "type": "object",
        "required": [
          "dict_type",
          "dict_name"
        ],
        "title": "DictTypeCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}